pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.3",
    "uvicorn[standard]>=0.44.0",
]
